requirements.txt
asgiref==3.8.1
Django==4.2.16
django-filter==25.1
djangorestframework==3.15.2
importlib_metadata==8.6.1
Markdown==3.7
sqlparse==0.5.3
typing_extensions==4.12.2
zipp==3.21.0
